+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- [#2171](https://github.com/teambit/bit/issues/2171) - fix ComponentNotFound when using `bit export` with no args and a flattened dependency was converted from no-scope to a remote-scope
 - add infrastructure for feature-toggle
 - deprecate files overrides (using file:// prefix)
 - disallow adding individual files by `bit add` unless `--allow-files` flag is used